bcache: Auxiliary search trees v3
This version removes the restriction that the auxiliary search trees
must be full; now we can construct trees with an arbitrary number of
nodes. This also lets us create trees with exactly one node per
cacheline in the set they're indexing.

To efficiently go from a node in the auxiliary search tree to the key it
points to, we now have code to convert from an index in the tree's array
to what it would be in an inorder traversal.

Change-Id: I86f5071b4f8becb38f05d67a384dd4a026469c79
Signed-off-by: Kent Overstreet <koverstreet@google.com>