bcache: Fix a data corruption bug in btree_sort_fixup()
When reading a btree node back in from disk, it'll have multiple sorted
sets of keys, and keys in different sets may overlap because they were
written at different times. Originally, readjusting the keys to not
overlap was done with the same code that handled key insertions, but
using it to check all the keys was roughly O(n^2*log(n)). So, awhile ago
the code was changed to readjust the overlapping extents during an n-way
mergesort.

The bug was that the heap of sets to be sorted didn't use the correct
comparison function; the way the code is written if two extents start at
the same offset they newer key has to be sorted first. The code handled
this, but only for the first two children and the parent in the heap;
i.e. I'd gotten it mostly right but it never clicked while I was writing
the code that that needed to be part of the comparison function. The bug
is painfully obvious when looking at the correct version, doing it right
is much cleaner.

Google-Bug-Id: 551185
Change-Id: I359cb01b0a9e0edf04fda30b3f2615dd72aa0cdc
Signed-off-by: Kent Overstreet <koverstreet@google.com>