bcache: Fix a cache coherency bug
If a cache miss races with a write, we could end up trying to add stale
data to the cache over the write we raced with.

This was previously handled by never overwriting good data when we're
inserting a key into the btree from a cache miss.

The bug was that if the write bypassed the cache, we'd insert a key with
no pointers (invalidating that region of the cache) but ptr_bad()
returns true for that key, thus we didn't guard against the race.

The correct way to do this is going to be to check that this spot in the
btree is the same as it was when the cache miss occured.  This is
actually what is done for background writeback. In the meantime though,
we can workaround by checking for !KEY_PTRS() in addition to !ptr_bad()
in check_old_keys(); this isn't rigorously correct but will work for
now.

Google-Bug-Id: 5562969
Change-Id: I1aebd9fc320c25b1746b96440121123eefff9892
Signed-off-by: Kent Overstreet <koverstreet@google.com>