bcache: Don't segregate dirty data
Using separate open buckets for dirty and non-dirty data was meant to be
an optimization - to reduce the percentage of buckets that are dirty -
but I don't think it was ever a great idea, and data is only dirty for a
bounded amount of time so it doesn't really matter. Taking it out
cleaned up the code a bit.

Also rename a few things for consistency

Change-Id: I408bec3b1b798a6669d692cd3907f80ecef7cf31
Signed-off-by: Kent Overstreet <koverstreet@google.com>