bcache: Fix a data corruption bug
When inserting data into the cache on cache miss, if we read into the
buffers from the original bio whoever was doing the original read (i.e.
userspace doing a dio read) might scribble over the data - or pass in an
iovec that all points to the same memory. Then, when we insert that data
into the cache we persist the corrupt data onto the SSD.

Fix this by reading into bounce buffers when we're going to insert the
data into the cache.

Change-Id: I9badc338a0928f587f3a93c3acd815a2c92b383e
Signed-off-by: Kent Overstreet <koverstreet@google.com>