bcache: Rename/cleanup btree memory cache stuff
Also moved some things around so as to require fewer extern symbols:
then changed it to not take bucket_lock when allocating at startup.
That's necessary in order to use GFP_KERNEL for the startup allocations
without lockdep complaining, and it's safe; then changed it to use
GFP_KERNEL and GFP_NOIO in the right places.

Change-Id: Ia62f6960dbccaa106bdc1abfbc6353ebc9be19c6