bcache: Don't initialize keylist until it's used
Profiling indicated zeroing out struct search was using a fair amount of
cpu. On cache hit we don't even use the keylist in struct search, so
this moves it to the end of the struct and only zeroes out the first
part of struct search, then doesn't initialize it until it's used - this
way we'll touch a few less cachelines.

Change-Id: I41e3a25bb1c4697b601b3221d1d838683af27973