bcache: Don't refcount cache hits
From profiling the atomic inc/dec of a bucket's refcount on cache hit
were taking significant cpu. So instead of doing that, we now check
after the IO completed if the ptr went stale in the meantime; if so, we
just handle it as if it was a read error and reread from the backing
device.

Change-Id: I71248384c14fb7cda07ae72afbc42a37a05f133d