bcache: Fix a race in the journalling error path
If bio_insert() didn't finish the first time through, we're going to
call bcache_journal() multiple times. If the second call happens before
the journal write from the first call finishes, we won't be able to wait
on the new journal write and we'd trigger the old BUG_ON(); fix this by
adding a closure_sync() - this waits for the first journal write if
there was one.

Change-Id: Ifb7224ef01b285e7319e0a356ac29601cb37631d