bcache: Fix btree_free() to cleanup properly
btree_free() previously assumed that if the node it was freeing had a
write set up that someone was explicitly waiting on, something had gone
wrong; it had BUG_ON(b->write->owner).

The BUG_ON() could trigger if btree_split() tried to free a btree node
immediately after a garbage collection, and that gc tried to write a
btree node that already had a write in flight.

However, the original assumption was wrong - btree_split() doesn't free
the original btree node until the new replacement node(s) are on disk.
Thus, if someone was waiting for the write we're freeing, that data has
already been written.

The fix is to just do the same cleanup that btree_write_endio() does, so
the patch consolidates that code.

Also, If we're freeing a non leaf btree node and it had keys that
weren't written yet, b->prio_blocked might be nonzero. Don't leak that
refcount.

Change-Id: I9597f36cb14824d6c5467db1130a3491890ce2e1
Signed-off-by: Kent Overstreet <koverstreet@google.com>