bcache: Fix auxiliary search trees without rebuilding
When we're inserting a new key into the btree, it may overlap with an
existing key in such a way that the old key has to change - thus
invalidating the auxiliary search tree.

Previously, we just rebuilt the entire tree, but this was far too
expensive if the workload hit it a lot - overwriting cached data with a
different blocksize would do it.

The new code fixes only the nodes in the auxiliary search tree that were
affected (possibly none), and is now roughly O(log log n) in the size of
the auxiliary search tree.

Change-Id: If97e9e0aaa557e6a3cd06ef105a2b5b35a9f2d70
Signed-off-by: Kent Overstreet <koverstreet@google.com>