bcache: Fix a deadlock in the journalling code
If we try to insert too many keys at once, the keylist may have more
keys than will fit in an empty journal write - and deadlock.

Prevent this from occuring by changing keylist_realloc(). It might be
conceptually better to have bcache_journal() handle that case, but that
would've been much more invasive and journalling a set of keys would no
longer be atomic, which might be a useful property in the future.

Change-Id: Iab71f1b21830732b4da5469354b138a874ceba42
Signed-off-by: Kent Overstreet <koverstreet@google.com>