bcache: Start writing to the journal in the correct spot
Previously, we always started writing journal entries at the start of
the journal; this meant we'd overwrite journal entries that were still
needed, and if the machine was shut down soon enough we'd be missing
journal entries next time we replayed the journal.

As an added bonus, this gets us support for discontiguous journals.

Change-Id: Ic45616ea397910d91b71582aa0f5fd50966e269a