bcache: Fix a pair of deadlocks
If a btree insert has to split and takes the try_harder lock, but
another thread locked the root of the btree and blocked on that lock
while the first thread was unlocking the root of the btree so it could
take a write lock, we'd deadlock. Fix this by not flipping on blocking
mode until it's actually needed.

The second deadlock was caused by garbage collection blocking in
btree_alloc() when allocating a new btree root - garbage collection
can't block on allocation.

Change-Id: Iae752b74f298d337e12aa6a770e055e7043ca89b