bcache: Refactor mca_bucket_alloc()
Previously, mca_bucket_alloc() only returned true the bucket it
allocated if it was able to allocate everything - we didn't return a
stub struct btree.

That made sense in the past, but this change will make more sense when
we're keeping around stub btree nodes, and simplifies alloc_bucket() a
bit.