bcache: move to drivers/md/
This is in preparation for being exposed as a md or dm device.  The
rationale is that userspace tooling already knows how to generically
manipulate these composite block devices, so there is not much
motivation to add new bcache-specific tooling.

As seen with raid there is precedent for md and dm to share a common
personality backend and separate the frontend virtual block device
presentation.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>