bcache: Moving garbage collector
The moving garbage collector works by finding buckets with some
dirty data and moving the dirty data into a new bucket.  If the
dirty data is associated with a flash only volume, this will free
buckets. As a bucket collects flash only data, it will stay around
for a while until some of the data is overwritten with new content
and enough of the bucket becomes reclaimable to make it worth moving
the dirty data so that the bucket can be reclaimed.