bcache: Fix an accounting integer overflow
We store the various statistics in long integers, but those are updated
on a timer from 32 bit atomic integers. If that 32 bit int wraps around
in the ~13 seconds between timer intervals, it'd end up adding a very
large negative number - since atomic_t is signed it'd get sign
into a signed long, then converted to unsigned long... not what we want.

Convert it to unsigned int first, so we get the full 32 bits and if it
wraps that we get 0 instead of ULONG_MAX.