bcache: track age of written data
This adds a write_prio to buckets to track the age of written
data. It is set to max on every write, and decremented every so
often after. It is soon to be used in the moving garbage
collector to group data of similar age (into so called
generations), while it's compacting data.

Signed-off-by: Nicholas Swenson <nks@daterainc.com>
Signed-off-by: Kent Overstreet <kmo@daterainc.com>