bcache: make MAP_END_KEY less error prone
Now, if you ask for MAP_END_KEY you get passed NULL for the key at the end of
each btree node - this seemed uglier to me at the time than what was done before
this patch, but the old behaviour led to a bunch of bugs in the new inode/dirent
code that's still out of tree. Now, if MAP_END_KEY is misused it'll be caught
right away with a null pointer deref.

Signed-off-by: Kent Overstreet <kmo@daterainc.com>