bcache: Add refcount to open_buckets
Prep work for finally getting rid of bucket->pin!

The way we're gonig to get rid of bucket pin is by having buckets owned by
something that the garbage collector can find until after we've inserted the new
keys that point to that bucket into the btree.

This adds code for allocating struct open_buckets, and freeing them when their
refcount goes to 0, and reworks bch_alloc_sectors() for the new way of
allocating buckets and to pass the pointer to the struct open_bucket back to
bch_data_insert_start().

Signed-off-by: Kent Overstreet <kmo@daterainc.com>