bcache: Massive journalling rework
Prep work for moving where keys are journalled to within the btree insertion
code: This adds bch_journal_write_get() and bch_journal_write_put(), which in
the next patch will be used from the btree code to journal keys immediately
after they've been added to a btree node, while that node is still locked.

This also does some general refactoring, and it also changes the journalling
code to not require a workqueue for anything important (in particular, the next
journal write if one needed to go out had to be kicked off from the system_wq).

This will help to avoid deadlocks when the journalling code is being used from
more interesting contexts.

Signed-off-by: Kent Overstreet <kmo@daterainc.com>