bcache: priorities algorith changed to CLOCK
The previous algorith always set read_prio to max, and then
decremented all bucket prios every so many bytes. This required
looping over all the buckets which could take awhile as cache sizes
get larger. The new algorith instead uses a circular clock, a u16 thats
incremented every so often, and sets read_prio to this clock when
read from. This doesn't require looping over all the buckets to increment
the clock every time.

Occasionally, because data may be in the cache for
an arbitrary amount of time, we may need to loop over all the buckets to
rescale the prios, and create more space for new prios.

Signed-off-by: Nicholas Swenson <nks@daterainc.com>