bcache: Make bucket allocation asynchronous again
Bucket allocation in bcache used to be asynchronous, ages and ages ago, but for
awhile I was making various asynchronous stuff synchronous.

With moving gc and now tiering becoming more important, and also the addition of
refcounted struct open_bucket to get rid of the bucket refcount, this has turned
out to be impractical - to avoid deadlocks we'd need a ridiculous number of
workqueues.

We can avoid the deadlocks by going back to making various things asynchronous -
the conversion is pretty straightforward.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>