bcache: add bch_bucket_wait()
Factor out the logic for adding a closure to the bucket_wait waitlist.
Add code to wake up btree GC thread if we are waiting on the tiering
or moving GC reserves to refill. In an upcoming patch, the allocator
will no longer wake btree GC to make progress, so instead tiering and
moving GC will need to do that on their own.

Change-Id: I5886c5e2fdba28818d5d30b0a954cdfb1f5bba4d