bcache: handle read bucket invalidate race in bch_read()
Previously there was a chance this would return wrong data. Now, we
use a custom endio function for bch_read()'s leaf bios. If the ptr
is stale upon completion of the read, requeue the bio on a
per-cache set list and kick off a work item to feed it back through
bch_bio().

Change-Id: Iba6570f354b605b3a5df1b243a2bf7828c823f2c