bcache: add a writeback error tracepoint
The dirty_io struct put the error field after the bio. However, we were
allocating the dirty_io together with the biovec, and assuming that
bio->bi_inline_vecs points to this area. As a result, the error field
was being clobbered with garbage, and we were not actually proceeding
to the write phase.

Add a writeback_error tracepoint to notice this condition in the future.

This fixes a regression from 7fa7d84a.

Change-Id: Ia7018b5509569539fdd7de24630067620a3f40db