bcache: fix a refcounting bug and return code
Fix a use-after-free bug -- if registering the cache or cached_dev fails,
we have already put our reference on the bdev because we put the kobject
which puts the bdev.

Differentiate a device that's already been registered with bcache from a
device that is busy by returning -EBUSY from register_bcache() in the
latter case.

Change-Id: I35bd164cebd7cb7eea2a06179e9f2c6505e2fcc7