bcache: fix a refcounting bug and return code
Fix a use-after-free bug -- if registering the cache or cached_dev fails,
we have already put our reference on the bdev because we put the kobject
which puts the bdev.

Differentiate a device that's already been registered with bcache from a
device that is busy by returning -EBUSY from register_bcache() in the
latter case.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>