bcache: keybuf key marking
When GC goes to mark writeback_keys we don't need to mark those sectors as
live; all we care about is that the gens don't wrap around if the buckets
are invalidated. Check for stale pointers in read_dirty_endio() to handle
this.

Also, we have to do this for tiering and moving GC keys too.

Change-Id: I19609f000aeba1952ecf3cb9ea81274a326fb093