bcache: redo GC marks using compare and exchange
This patch changes the representation of bucket marks to an array of a new
bucket_mark struct. Elements of this array are updated atomically with a
new bucket_cmpxchg() macro. The old GC mark states map roughly as follows:

- GC_MARK_METADATA => mark.is_metadata = 1
- GC_MARK_RECLAIMABLE => mark.dirty_sectors = 0, mark.cached_sectors > 0
- GC_MARK_DIRTY => mark.dirty_sectors > 0, or mark.owned_by_allocator = 1
- GC_SECTOR_COUNT => mark.dirty_sectors, or mark.cached_sectors

Cached and dirty sectors are now counted separately. Also, a new
mark.owned_by_allocator flag is used for buckets that are on free lists,
and open buckets, in place of GC_MARK_DIRTY.

Right now, the owned_by_allocator flag is not cleared until the next
garbage collection. This ensures that open buckets do not become
candidates for invalidation, since they do not have an up to date sector
counts. Once sector counts are kept up to date by foreground writes, the
owned_by_allocator flag can be cleared once the open bucket is put.

Change-Id: I244af7e3685f28424570c69599043de1c94e9de7