bcache: don't foul up PTR_BUCKET in bch_cut_front()
If the key points to the last extent in a bucket and we shorten it to
zero length, we end up changing the key to point to the next bucket.
Normally we never look at pointers of zero-length keys, but once we
are updating bucket sector counts in the foreground write path, we
will end up looking at buckets pointed at by keys that have potentially
just keep shortened to zero length.

Signed-off-by: Kent Overstreet <kent.overstreet@gmail.com>