discards: functional rewrite
Discards now use a custom bcache map function instead of using the
data_insert path. This is to prevent extraneous key insertions when
discarding large portions of the extent space, which may or may not
have be actually allocated. The data_invalidate function (of the
data_insert_path) would insert delete keys to cover the whole range
specific, regardless if keys existed in that range or not. The new
map function only inserts delete keys when a key is actually found.
Also, the new function inserts the largest keys possible, so each
insertion covers as much of the range possible. This minimizes the
amount of insertions (and node writes) done.

Change-Id: I2fe1bb7ba1b1a5bbd1951763f27ac95be915313a