bcache: don't count buckets twice
If owned_by_allocator is set, we don't want to count the bucket
as metadata, dirty or cached since it is already counted as alloc.

Similarly, if dirty_sectors is set, don't count the bucket as
cached.

This was causing buckets_free_cache() to return a negative value,
causing btree node allocations to fail while splitting interior
nodes, which is not supposed to happen.

This fixes a regression from "refactor bucket stats code".

Also, fix a misleading variable name in alloc.c, and fix argument
order for btree_check_reserve_fail tracepoint.

Change-Id: I9e52b9008658bc35688033117f07ee6820ebfc8f