bcache-ioctl: refactoring COPY ioctl
Created a generic IO interfact anyone can use. Refactored the algorthm
to make it a bit more clear as to whats going on. Also fixed it so it
works for arbitrary ranges. Previously, it would loop infinitely if the
copy range extended beyond that last found extent within that range.

Change-Id: I9478b953e3e07778645d38640332e01ef3fa61dd